fix(#227): validate client durations at the boundary; widen presence counters
Round-2 findings fixed at the ROOT (boundary validation), not per-site.

Boundary validation (finding 2, root-cause fix):
- New MAX_DURATION_NS domain (i64::MAX/4 ~= 73 years) + a single
  validate_duration_ns() funnel in plan.rs. Every client-controlled
  duration entering the range/sliding engine -- the [range] selector and
  the request step -- is validated ONCE there into 1..=MAX_DURATION_NS and
  carried as a plain i64. Out-of-domain values are a clean named 400
  (ReadError::DurationOutOfRange, mapped in logs_api + prom_api).
- MetricPlan.range_ns and ClientWindow.range_ns are now i64, so the flagged
  `window.range_ns as i64` narrowing is GONE rather than made saturating.
  Verified the pre-fix hazard was real: u64 i64::MAX+1 narrowed to
  i64::MIN, inverting the window.
- The leafless tree (vector(n) / 2+2) has no metric leaf, so the same
  validation is applied in plan_metric_expr -- a hostile step can no
  longer reach materialize_vector_lit's grid math.

Counter width proof (finding 1):
- present_cover/running widened i32 -> i64 with the bound stated in the
  doc comment: |counter| <= surviving collision groups <= scanned rows <=
  LOGQL_SCAN_BUDGET_BYTES_CEILING (one byte per row minimum) < i64::MAX.
  Overflow is impossible by construction, so no cap/checked arithmetic is
  needed; i32 (2.1e9) was genuinely reachable under a multi-GiB budget.

Tests (closing the round-2 gaps):
- AC8 extended to the MUTATING path: MutGroup.int_cells AND pt_cells both
  charge retention exactly once per created cell/retained point, released
  on finish.
- Hostile-duration rejection at the planner boundary (unit + end-to-end
  through plan(), incl. the leafless tree) for 0, >MAX_DURATION_NS,
  i64::MAX+1 and u64::MAX.
- Presence-counter overflow proof pinned (exact past i32::MAX, budget
  ceiling < i64::MAX).
- The extreme-window golden now uses MAX_DURATION_NS as the in-domain
  ceiling (a larger step is rejected earlier, by the boundary).

Guards tripped for real (disabled the validation -> hostile test failed;
demonstrated the i64::MIN wrap) then reverted.

Gauntlet: pulsus-read+pulsus-server 1798 passed 0 failed; clippy
--all-targets -D warnings clean; fmt clean. Live CH re-verified:
explain_indexes 25, live_metrics_engine 25, rollup_differential 8.

+++ crates/pulsus-read/tests/sql_snapshots.rs
@@ -553,7 +553,7 @@ fn expected_sliding_reason() -> String {
 
 /// The `[5m]` selector range in nanoseconds — issue #227 makes it the
 /// `rate`/`bytes_rate` divisor AND the sliding window width (never `step`).
-const RANGE_NS: u64 = 300_000_000_000;
+const RANGE_NS: i64 = 300_000_000_000;
 
 #[test]
 fn rate_range_slides_raw_and_divides_by_the_range() {
@@ -567,7 +567,7 @@ fn rate_range_slides_raw_and_divides_by_the_range() {
     assert_eq!(mp.table, "log_samples");
     // `rate` divides by the `[range]` (5m), NOT `step` — the
     // `rate([1m]) ≠ rate([10m])` fix.
-    assert_eq!(mp.rate_window_ns, Some(RANGE_NS));
+    assert_eq!(mp.rate_window_ns, Some(RANGE_NS as u64));
     assert_eq!(mp.routing.chosen, pulsus_read::logql::RouteChoice::Raw);
     assert_eq!(mp.routing.reason, expected_sliding_reason());
 }
@@ -592,7 +592,7 @@ fn bytes_rate_range_slides_raw_and_divides_by_the_range() {
     );
     assert!(!mp.rollup);
     assert!(mp.client.is_some());
-    assert_eq!(mp.rate_window_ns, Some(RANGE_NS));
+    assert_eq!(mp.rate_window_ns, Some(RANGE_NS as u64));
     assert_eq!(mp.routing.reason, expected_sliding_reason());
 }
 
@@ -882,7 +882,7 @@ fn range_metric_spec_carries_the_caller_supplied_step() {
     // Issue #227: the emit grid starts at the caller's `start`; the SCAN
     // lower bound is widened back by the `[5m]` range for the lookback.
     assert_eq!(mp.grid_start_ns, START_NS);
-    assert_eq!(mp.start_ns, START_NS - RANGE_NS as i64);
+    assert_eq!(mp.start_ns, START_NS - RANGE_NS);
     assert_eq!(mp.range_ns, RANGE_NS);
     assert_eq!(mp.end_ns, END_NS);
 }
